requirements.txt
addict
blend-modes
segment-anything
pyqt5
yapf
platformdirs
translate
pycryptodome
pycryptodomex
ntplib
py-cpuinfo  
transparent_background
requests
torch
torchvision
Pillow
numpy
scipy
scikit-image
opencv-python
aiohttp
pilgram
customtkinter